requirements.txt
confluent-kafka==1.4.2
python-decouple==3.3
tweepy==3.8.0
